feat(frontend): Add running runs on the script detail page
