+++ frontend/src/lib/utils/JobManager.ts
@@ -1,0 +1,9 @@
+import type { CompletedJob, Job } from '$lib/gen'
+
+class JobManager {
+	private jobs: Job[] | undefined
+
+	constructor() {
+		// Initialize variables if needed
+	}
+}
